refactor(import): extract CallMapper; AegisXmlParser is now a ~200-line orchestrator
Final mapper (#49). Move the top-level calls row insert/update (and xmlToArray)
into NwsCad\Import\Mappers\CallMapper, returning the db id. insertCall() is now
a thin orchestrator: CallMapper then the eight child mappers. The dead
parseDateTime/parseBoolean/parseInt/xmlToArray wrappers are removed.

Also drop the obsolete DateTimeParsingCharacterizationTest: it reflected on the
now-removed AegisXmlParser::parseDateTime() wrapper; its golden-value coverage
is fully served by DateTimeParserTest against the extracted DateTimeParser.

AegisXmlParser: ~1337 lines at #49's start -> 223 lines.

Behavior-preserving: full Unit (320) + Integration (184) + Security (61) suites
green on MySQL; parser coverage set green under pcov (37 tests, no risky);
PostgreSQL characterization green (6 tests/125 assertions).

Part of #49.
