+++ tests/Integration/ImporterCharacterizationTest.php
@@ -32,6 +32,7 @@ use PHPUnit\Framework\TestCase;
  * @uses \NwsCad\Import\Mappers\IncidentMapper
  * @uses \NwsCad\Import\Mappers\NarrativeMapper
  * @uses \NwsCad\Import\Mappers\UnitMapper
+ * @uses \NwsCad\Import\Mappers\CallMapper
  * @uses \NwsCad\Config
  * @uses \NwsCad\Database
  * @uses \NwsCad\FilenameParser
